# back/public/swagger.json
{"openapi":"3.0.0","paths":{"/api/auth/signin":{"post":{"operationId":"AuthController_signin","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SigninInput"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"tags":["auth"]}},"/api/auth/signup":{"post":{"operationId":"AuthController_signup","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserInput"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"tags":["auth"]}},"/api/auth/verifyOtp":{"post":{"operationId":"AuthController_verifyOtp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyOtpInput"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyOtpResponse"}}}}},"tags":["auth"]}},"/api/users/me":{"get":{"operationId":"UsersController_me","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"tags":["users"]}}},"info":{"title":"Novity Slack API","description":"Details of Novity Slack API","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"SigninInput":{"type":"object","properties":{"email":{"type":"string"}},"required":["email"]},"UserResponse":{"type":"object","properties":{"id":{"type":"number"},"email":{"type":"string"},"name":{"type":"string"}},"required":["id","email","name"]},"CreateUserInput":{"type":"object","properties":{"email":{"type":"string"},"name":{"type":"string"}},"required":["email","name"]},"VerifyOtpInput":{"type":"object","properties":{"otp":{"type":"string"},"email":{"type":"string"}},"required":["otp","email"]},"VerifyOtpResponse":{"type":"object","properties":{"token":{"type":"string"},"refreshToken":{"type":"string"},"data":{"$ref":"#/components/schemas/UserResponse"}},"required":["token","refreshToken","data"]}}}}